requirements.txt
streamlit
requests
gspread
google-auth
pandas